+++ CHANGELOG.md
@@ -38,6 +38,7 @@
 + [x] Elements colors: optimize less for using `each` function
 + [x] Additional colors: optimize less for using `each` function
 + [x] Element: add class `.accent-block` for using with accent colors classes `primary`, `alert`, `info`, `warning`, ...
++ [x] Dialog: add close button to top left corner with attribute `data-close-button`
 
 ### 4.2.39
 + [x] NavView: minor improve styling